код распределен на модули
