+++ js/until.js
@@ -1,0 +1,34 @@
+const comments = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+const names = [
+  'Виктор Пупов',
+  'Геннадий Изюмов',
+  'Оксана Тетёхина',
+  'Ваня Комментаторов',
+  'Фотоспец3002'
+];
+
+const descriptions = [
+  'здесь мог быть lorem, но ты просто чилишь',
+  'Русалка села на шпагат',
+  'Колобок повесился'
+];
+
+const photoValues = {
+  PHOTOS_COUNT: 25,
+  RANDOM_MAX_NUMBER_AVATAR: 6,
+  RANDOM_MIN_NUMBER_AVATAR: 1,
+  RANDOM_MAX_NUMBER_COMMENTS: 30,
+  RANDOM_MIN_NUMBER_COMMENTS: 0,
+  RANDOM_MAX_NUMBER_LIKES: 200,
+  RANDOM_MIN_NUMBER_LIKES: 15,
+};
+
+export {comments, photoValues, names, descriptions};
